fix(link): avoid passing in animateUnderline as attribute (#3787)
Avoid passing in animateUnderline as attribute. Otherwise results in warning:
```
Warning: React does not recognize the `animateUnderline` prop on a DOM element. If you intentionally want it to appear in the DOM as a custom attribute, spell it as lowercase `animateunderline` instead. If you accidentally passed it from a parent component, remove it from the DOM element.
```

Test Plan: Create a Link with animateUnderline prop and no longer see the warning.