fix(select): handleBlur should check event type before bound check (#4960)
When we handle a blur event in the select-component, we want to check if
the event is transferring focus to an element which is a child of the
select component before we un-focus.

Because handle blur takes FocusEvents or MouseEvents, there is a hole in
the logic, where a FocusEvent which does not have a relatedTarget (which
is what happens if you imperatively blur with element.blur() or via the
controlRef.setInputBlur() ) will pass to the second conditional check in
handleBlur, which should only be applied to mouse events (i.e. If you
click on a new element, we check if that new element (event.target) is
inside our anchor).

In the case of a FocusEvent, event.target is the element being blurred
(the input) so it will always be inside our anchor.